example/requirements.txt
django>=4.2
djangorestframework>=3.14
drf-spectacular>=0.27
drf-mcp-docs
